Fix (Custom Sound or Icon not Showing in Release)
People are coming up with the Issue of not seeing the custom icons or the custom sounds in the Action (i.e in the APK or release). So I Decided to Search up the whole web, For this, and Found that, Adding just Two lines of code in the AndroidManifest.xml we can remove the error of not seeing the Custom Sounds and Icons in the Action.